+++ adventofcode-2021.tests/Day1Tests.cs
@@ -12,7 +12,7 @@ namespace adventofcode_2021.tests
             var input = new string[] { "199", "200", "208", "210", "200", "207", "240", "269", "260", "263" };
 
             var day1 = new Day1();
-            int result = day1.Part1(input);
+            int result = (int)day1.Part1(input);
 
             result.Should().Be(7);
         }
@@ -23,7 +23,7 @@ namespace adventofcode_2021.tests
             var input = new string[] { "199", "200", "208", "210", "200", "207", "240", "269", "260", "263" };
 
             var day1 = new Day1();
-            int result = day1.Part2(input);
+            int result = (int)day1.Part2(input);
 
             result.Should().Be(5);
         }
